Refactoring ongoing, added random list operator
As part of the on course efforts to clean and improve the current
codebase, major changes are taking place.

A cost change function is introduced, that calculates the total cost
of an alternate route in constant 8 lookups. implementation in progress.

The major change that has taken place until now, is the introduction of an
operator to reduce boilerplate code in operations that require heavy generation
of random values of the same type. '~#' can be used to generate a list of such
values, by taking care of the resultant StdGen.