Set up database access, refactored datatype decls
In this commit, access to the database was properly set up within the
request handlers of the servant app. it is now possible to liftIO a
database action inside the monad. It is now clear how to return an HTTP
error message as well.

The datatype declarations were moved into their own file, Yesod style.

general refactoring and cleanup was also performed.