completed and refactored handlers, extended schema
In this commit, effort was directed towards completing the first sketch
of an API, with the first two objects to be CRUDed implemented (the Account
object and the Vehicle). Future developments should see the completion of
the API by including all objects.

The handlers were at the same time placed on their own modules, specializing
the purpose of the Exec module (that of being the entry point of the app).
A new module, Import, was created to reduce boilerplate.

Finally, the database schema was updated to generate the first draft of what
will eventually become the definite version.