Finished basic API, moved to package.yaml
The basic REST services for the current datatypes has been completed.
It's possible now to add and read entries from the database, one at a time
or (in the case of viewing) all of them.

The package information has been moved the package.yaml file. The .cabal
file is no longer being manually modified.